reformat conv_layer via functional
