+++ conv_layer.py
@@ -4,6 +4,7 @@ Convolution layers
 
 import numpy as np
 from layer import Layer
+import functional as F
 import multiprocessing as mp
 # This is for windows
 from multiprocessing.pool import ThreadPool as Pool
@@ -13,7 +14,6 @@ from multiprocessing.pool import ThreadPool as Pool
 # 2. stride dilation calculation when doing full-convolution
 # 3. grouping 
 # 4. check multiprocessing
-
 
 
 class Conv(Layer):
@@ -29,16 +29,15 @@ class Conv(Layer):
             bias=False
     ) -> None:
         
-        if isinstance(kernel_size, int):
-            # case when kernel size is int for square kernel
-            self.filters = np.random.randn(in_channels, out_channels, kernel_size, kernel_size)
-        else:
-            # case when kernel size is tuple
-            self.filters = np.random.randn(in_channels, out_channels, kernel_size[0], kernel_size[1]) 
+        # reformat paramters into tuples
+        kernel_size, self.stride, self.padding, self.dilation = F.paramtuple(
+            kernel_size, stride, padding, dilation
+        )
+        
+        # initalize filters
+        self.filters = np.random.randn(in_channels, out_channels, kernel_size[0], kernel_size[1])
 
-        self.stride = stride
-        self.padding = padding
-        self.dilation = dilation
+        ### add grouping ###
         self.groups = groups
         
         # introduce bias for smaller network
@@ -56,9 +55,9 @@ class Conv(Layer):
         self.output = self.convntom(
             input,
             self.filters,
+            self.stride,
             self.padding,
-            self.dilation,
-            self.stride
+            self.dilation
         )
 
         # initialize bias at the very first forward pass 
@@ -88,9 +87,9 @@ class Conv(Layer):
                 filter_input.append((
                     self.input[0, ic], 
                     output_error[0, oc],
+                    self.stride,
                     self.padding,
-                    self.dilation,
-                    self.stride
+                    self.dilation
                 ))
 
         with Pool(mp.cpu_count()) as p:
@@ -110,10 +109,10 @@ class Conv(Layer):
         input_error = self.convntom(
             output_error, 
             self.filters,
-            padding=self.padding,
-            dilation=self.dilation,
             ### check stride calculation ###
             stride=self.stride,
+            padding=self.padding,
+            dilation=self.dilation,
             full=True
         )
 
@@ -123,181 +122,6 @@ class Conv(Layer):
 
         return input_error
 
-
-    def pad(self, data, padding):
-        """
-        Helper function to add padding before convolution
-        """
-        
-        if isinstance(padding, int):
-            # case when padding is int for square kernel
-            row_pad = col_pad = padding
-        else:
-            # case when padding is tuple
-            row_pad, col_pad = padding
-
-        ### Look up online to optimize ###
-
-        data = np.pad(data, ((row_pad, row_pad), (col_pad, col_pad)), mode="constant")
-
-        return data
-
-
-    def mult(
-            self,
-            data, 
-            filter,
-            r,
-            c,
-            out_r, 
-            out_c
-    ):
-        """
-        Helper function for 1 mult-operation in convolve.
-        """
-
-        data_r = out_r * self.stride + r * self.dilation
-        data_c = out_c * self.stride + c * self.dilation
-
-        return filter[r, c] * data[data_r, data_c]
-
-
-    def convolve(
-            self, 
-            data, 
-            filter,
-            out_r, 
-            out_c
-    ):
-        """
-        This is the function calculating 1 convolution process.
-        """
-        
-        f_H, f_W = filter.shape
-        input = []
-        
-        for r in range(f_H):
-            for c in range(f_W):
-                input.append((data, filter, r, c, out_r, out_c))
-        
-        with Pool(mp.cpu_count()) as p:
-            output = p.starmap(self.mult, input)
-            p.terminate()
-            p.join()
-
-        return sum(output)
-    
-
-    def conv1to1(
-            self, 
-            data, 
-            filter,
-            padding,
-            dilation,
-            stride,
-            full=False
-    ):
-        """
-        This is the function calculating 1 unit data to 1 filter,
-        for later integration into convNtoM.
-        Hand-select data and filter to be process.
-        Take padding, dilation, stride... into consideration.
-        """
-        
-        # if full-convolution, flip the filter then switch data and filter
-        # since the calculation is:
-        # full-convolution(filter, output_error)
-        # also calculate padding
-        if full:
-            filter = np.flip(filter, axis=0)
-            filter = np.flip(filter, axis=1)
-
-            padding = (
-                np.abs(filter.shape[0] - 1), np.abs(filter.shape[1] - 1)
-            )
-
-        # pad data if needed
-        if isinstance(padding, int):
-            if padding > 0:
-                data = self.pad(data, padding)
-            padding = (padding, padding)
-        else:
-            data = self.pad(data, padding)
-
-        in_H, in_W = data.shape
-        f_H, f_W = filter.shape
-
-        # if full-convolution, out_H and out_W would match the input shape
-        if full:
-            out_H, out_W = self.input.shape[2], self.input.shape[3]
-        else:
-            out_H = int((in_H + 2 * padding[0] - dilation * (f_H - 1) - 1) / stride + 1)
-            out_W = int((in_W + 2 * padding[1] - dilation * (f_W - 1) - 1) / stride + 1)
-        input = []
-
-        for r in range(0, out_H):
-            for c in range(0, out_W):
-                input.append((data, filter, r, c))
-        
-        with Pool(mp.cpu_count()) as p:
-            out_data = p.starmap(self.convolve, input)
-            p.terminate()
-            p.join()
-        
-        out_data = np.array(out_data)
-
-        return np.reshape(out_data, [out_H, out_W])
-    
-
-    def convntom(
-            self, 
-            data, 
-            filters,
-            padding,
-            dilation,
-            stride,
-            full=False
-    ):
-        """
-        This is the function calculating convolution for 
-        n input channels to m output channels.
-        The input data shape is expected as: 
-        (1, in_channel, in_H, in_W)
-        The filters shape is expected as:
-        (in_channel, out_channel, f_H, f_W)
-        The output shape is expected as:
-        (1, out_channel, out_H, out_W)
-        """
-
-        # if full-convolution, flip in/out_channel
-        if full:
-            filters = np.swapaxes(filters, 0, 1)
-
-        input = []
-        in_channels, out_channels = filters.shape[0], filters.shape[1]
-
-        for oc in range(out_channels):
-            for ic in range(in_channels):
-                input.append((
-                    data[0, ic], 
-                    filters[ic, oc],
-                    padding,
-                    dilation,
-                    stride,
-                    full
-                ))
-        
-        with Pool(mp.cpu_count()) as p:
-            out_data = p.starmap(self.conv1to1, input)
-            p.terminate()
-            p.join()
-        
-        out_H, out_W = out_data[0].shape[0], out_data[0].shape[1]
-        # sum up result from in_channels to one combined result
-        out_data = np.sum(np.reshape(np.array(out_data), 
-                                     [out_channels, in_channels, out_H, out_W]), axis=1)
-        
-        return np.expand_dims(out_data, axis=0)
 
 
 if __name__ == "__main__":
